MNT: update to more reliable method of creating legends
This fixes a bug identified by @fperez

When calling `ax.legend` with one arg the list of strings is zipped
with the available artists in the Axes.  This is brittle because it
assumes the contents and order of this list. In Matplotlib 1.5.1 we
added a legend handler for the artists that are used to draw the
fill_between regions which caused them to be included in the list of
artists which will go in the legend and due to internal details about
how Matplotlib store's the children artists of the Axes the
fill_between artists are listed before the errorbar artists and thus
the legends end up shifted.

The primary change in this commit is to pass the correct label into
the plotting calls via the `label=` kwarg and to call `legend` with no
args.